text_interview/requirements.txt
livekit-agents[openai]>=1.2
livekit-api>=0.6
supabase>=2.6
python-dotenv>=1.0
